reloc: rename priors.py to utils.py
